alter to uint 64

+++ keccakhash.c
@@ -23,7 +23,7 @@ void keccak_hash(const char* input, char* output)
 
     
     //these uint512 in the c++ source of the client are backed by an array of uint32
-    uint32_t hashA[16], hashB[16];
+    uint64_t hashA[16], hashB[16];
     
     //uint256_t hashA[16], hashB[16];
